fix: refactored edit modal and implemented opening modal using redux store
