--- src/components/BtnCancelTrans/BtnCancelTrans.module.scss
@@ -1,48 +1,0 @@
-@import "./../../../public/modern-normalize.scss";
-@import "../../../stylesheet/vars.scss";
-@import "//stylesheet/fonts.scss";
-
-// MOBILE
-// TABLET
-// DESKTOP
-.btn {
-  margin-bottom: 37px;
-
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  
-  width: 300px;
-  height: 50px;
-  flex-shrink: 0;
-
-  border: 1px solid $color-brand-primary;
-  border-radius: 20px;
-  background: $background-light;
-
-  cursor: pointer;
-
-  &__label {
-    width: 164px;
-    height: 24px;
-    flex-shrink: 0;
-
-    color: $color-brand-primary;
-    text-align: center;
-    font-size: 18px;
-    font-style: normal;
-    font-weight: 400;
-    line-height: normal;
-    letter-spacing: 1.8px;
-    text-transform: uppercase;
-    @font-face {
-      font-family: 'circe-regular';
-      src: url('../../assets/fonts/circe-regular.otf');
-    }
-  }
-}
-
-.btn:hover {
-  box-shadow: 0px 4px 4px 0px rgba(0, 0, 0, 0.25);
-  transition: transform 0.3s ease;
-}
