+++ src/components/ButtonAdd/ButtonAdd.module.scss
@@ -1,51 +1,20 @@
+@import "../../../stylesheet/vars.scss";
 
-.btnDiv {
-
-    
-    display: flex;
-    justify-content: end;
-
-    z-index: 9;
+.buttonAdd {
+  cursor: pointer;
+  border-radius: 50px;
+  position: fixed;
+  bottom: 20px;
+  right: 20px;
+  z-index: 2;
+  @include tablet {
+    position: fixed;
+    bottom: 8px;
+    right: 8px;
+    z-index: 2;
+  }
+  @include desktop {
+    bottom: 40px;
+    right: 40px;
+  }
 }
-.btn {
-    cursor: pointer;
-    // transition: transform 0.3s ease;
-    // box-shadow: 10px 10px ;
-
-    // width: 74px;
-
-
-}
-
-.btn:hover {
-    // width: 70px;
-    // box-shadow: 10px 10px ;
-    // transform: rotate(45deg);
-}
-
-
-// .button {
-//     width: 44px;
-//     height: 44px;
-
-//     background-color: #24CCA7;
-
-//     border-radius: 50%;
-//     border: none;
-
-//     color: #fff;
-//     font-size: 36px;
-// }
-
-// .btn {
-//     width: 44px;
-//     height: 44px;
-// }
-
-// .btn2 {
-//     width: 44px ;
-//     height: 44px;
-//     border: none;
-//     // border-radius: 50%;
-// }
-
