+++ src/components/FormEdit/FormEdit.module.scss
@@ -1,0 +1,58 @@
+@import "./../../../public/modern-normalize.scss";
+@import "../../../stylesheet/vars.scss";
+
+.typeOfTransHead {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  margin-top: 24px;
+  margin-bottom: 24px;
+}
+
+.slashLabel {
+  margin: 0 11px 0 11px;
+}
+
+.form {
+  margin: 0;
+  padding: 0 11px 0 11px;
+  display: flex;
+  flex-direction: column;
+}
+
+.input {
+  margin-left: 40px;
+  width: 190px;
+  height: 24px;
+  border: none;
+  outline: none;
+
+  &__comment {
+    height: 52px;
+  }
+}
+
+.calendarWrap {
+  display: flex;
+  align-items: center;
+  margin-left: 40px;
+}
+
+.separatorShort,
+.separatorLong {
+  margin-top: 8px;
+  margin-bottom: 40px;
+  width: 280px;
+  height: 0;
+  flex-shrink: 0;
+  border-top: 1px solid #e0e0e0;
+}
+
+.calendarIcon {
+  position: fixed;
+  left: 240px;
+  width: 24px;
+  height: 24px;
+  fill: #4a56e2;
+  cursor: pointer;
+}
