+++ src/components/FormExpense/FormExpense.jsx
@@ -239,147 +239,3 @@ const FormExpense = () => {
 };
 
 export default FormExpense;
-
-// import React, { useRef } from "react";
-// import Select from 'react-select'
-
-// import styles from "./FormExpense.module.scss";
-
-// import { Formik, Field, Form, ErrorMessage } from "formik";
-// import * as Yup from "yup";
-// import moment from "moment";
-
-// import BtnAddTrans from "../BtnAddTrans/BtnAddTrans";
-// import Calendar from "../Calendar/Calendar";
-// import { ReactComponent as CalendarIcon } from "../../img/calendar.svg";
-// import transactionsApiSlice from "../../redux/slices/api/transactions/transactionsApiSlice";
-
-// const validationSchema = Yup.object().shape({
-//   category: Yup.string().required("Category is required"),
-//   amount: Yup.number()
-//     .typeError("Amount must be a number")
-//     .positive("Amount must be positive")
-//     .max(9999999, "Amount too large")
-//     .required("Amount is required"),
-//   date: Yup.date()
-//     .max(moment(), "Date cannot be in the future")
-//     .required("Date is required"),
-//   comment: Yup.string(),
-// });
-
-// const FormExpense = () => {
-//   const [addTransaction] = transactionsApiSlice.useAddTransactionMutation();
-//   const calendarRef = useRef(null);
-
-//   const openCalendar = () => {
-//     if (calendarRef.current && calendarRef.current.openCalendar) {
-//       calendarRef.current.openCalendar();
-//     }
-//   };
-
-//   return (
-//     <Formik
-//       initialValues={{
-//         category: "",
-//         amount: "",
-//         date: moment().format("YYYY-MM-DD"),
-//         comment: "",
-//       }}
-//       validationSchema={validationSchema}
-//       onSubmit={(values, { setSubmitting, resetForm }) => {
-//         addTransaction({
-//           income: false,
-//           amount: values.amount,
-//           category: values.category,
-//           date: values.date,
-//           comment: values.comment,
-//         });
-//         setSubmitting(false);
-//         resetForm();
-//       }}
-//     >
-//       {({ isSubmitting, setFieldValue, values }) => (
-//         <Form className={styles.form}>
-//           <div>
-//             <Field className={styles.input} as="select" name="category">
-//               <option className={styles.input__option} value="" label="Select category" />
-//               <option value="Products">Products</option>
-//               <option value="Main expenses">Main expenses</option>
-//               <option value="Car">Car</option>
-//               <option value="Self care">Self care</option>
-//               <option value="Child care">Child care</option>
-//               <option value="Household products">Household products</option>
-//               <option value="Education">Education</option>
-//               <option value="Leisure">Leisure</option>
-//               <option value="Other expenses">Other expenses</option>
-//               <option value="Entertainment">Entertainment</option>
-//             </Field>
-//             <ErrorMessage name="category" />
-//           </div>
-
-//           <div className={styles.separatorShort}></div>
-
-//           <div>
-//             <Field
-//               className={styles.input}
-//               name="amount"
-//               type="number"
-//               placeholder="0.00"
-//               autoComplete="off"
-//             />
-//             <ErrorMessage name="amount" />
-//           </div>
-
-//           <div className={styles.separatorShort}></div>
-
-//           <div>
-//             <div className={styles.calendarWrap}>
-//               <Calendar
-//                 ref={calendarRef}
-//                 value={values.date}
-//                 onChange={(date) =>
-//                   setFieldValue("date", date.format("YYYY-MM-DD HH:mm:ss"))
-//                 }
-//                 name="date"
-//               />
-
-//               <CalendarIcon
-//                 onClick={openCalendar}
-//                 className={styles.calendarIcon}
-//               />
-//             </div>
-//             <ErrorMessage name="date" />
-//           </div>
-
-//           <div className={styles.separatorShort}></div>
-
-//           <div>
-//             <Field
-//               as="textarea"
-//               className={`${styles.input} ${styles.input__comment}`}
-//               name="comment"
-//               placeholder="Comment"
-//               autoComplete="off"
-//             />
-//             <ErrorMessage name="comment" />
-//           </div>
-
-//           <div className={styles.separatorLong}></div>
-
-//           <BtnAddTrans
-//             onSubmit={
-//               isSubmitting
-//                 ? null
-//                 : () =>
-//                     document
-//                       .querySelector("form")
-//                       .dispatchEvent(new Event("submit"))
-//             }
-//           />
-//         </Form>
-//       )}
-//     </Formik>
-//   );
-// };
-
-// export default FormExpense;
