+++ src/components/TransactionListItem/TransactionListItem.jsx
@@ -1,17 +1,14 @@
-import React, { useState } from "react";
-import { ReactComponent as Pencil } from "./../../img/pencil.svg";
 import styles from "./TransactionListItem.module.scss";
-import transactionsApiSlice from "../../redux/slices/api/transactions/transactionsApiSlice";
+import React from "react";
+import { useDispatch } from "react-redux";
+import { ReactComponent as Pencil } from "./../../img/pencil.svg";
+import { openModal } from "../../redux/slices/modal/modalSlice";
 import EditTransaction from "../EditTransaction/EditTransaction";
+import transactionsApiSlice from "../../redux/slices/api/transactions/transactionsApiSlice";
 
-const TransactionListItem = ({
-  date,
-  category,
-  income,
-  comment,
-  amount,
-  _id,
-}) => {
+const TransactionListItem = ({ transaction }) => {
+  const { _id, date, category, comment, amount, income } = transaction;
+  const dispatch = useDispatch();
   const [deleteTransaction, { isLoading: isDeleting }] =
     transactionsApiSlice.useDeleteTransactionMutation();
 
@@ -19,58 +16,47 @@ const TransactionListItem = ({
     await deleteTransaction(_id);
   };
 
-  const [showEdittrans, setShowEditTrans] = useState(false);
-
-  const openModal = () => {
-    setShowEditTrans(true);
-  };
-
-  const closeModal = () => {
-    setShowEditTrans(false);
-  };
-
   return (
-    <ul className={styles.transaction}>
-      <li className={styles.transaction__item}>
-        <h4 className={styles.transaction__header}>Date</h4>
-        <p className={styles.transaction__data}>{date}</p>
-      </li>
-      <li className={styles.transaction__item}>
-        <h4 className={styles.transaction__header}>Type</h4>
-        <p className={styles.transaction__data}>
-          {income ? "Income" : "Expense"}
-        </p>
-      </li>
-      <li className={styles.transaction__item}>
-        <h4 className={styles.transaction__header}>Category</h4>
-        <p className={styles.transaction__data}>{category}</p>
-      </li>
-      <li className={styles.transaction__item}>
-        <h4 className={styles.transaction__header}>Comment</h4>
-        <p className={styles.transaction__data}>{comment}</p>
-      </li>
-      <li className={styles.transaction__item}>
-        <h4 className={styles.transaction__header}>Sum</h4>
-        <p className={styles.transaction__data}>{amount}</p>
-      </li>
-      <li className={styles.transaction__item__last}>
-        {/* TODO: add some spinner */}
-        <button className={styles.btn__delete} onClick={handleDelete}>
-          {isDeleting ? "Deleting..." : "Delete"}
-        </button>
-        <button className={styles.btn__edit} onClick={openModal}>
-          <Pencil className={styles.btn__edit__icon} />
-          Edit
-        </button>
-      </li>
-      {showEdittrans && (
-        <EditTransaction
-          onClose={closeModal}
-          income={income}
-          transactionId={_id}
-        />
-      )}
-    </ul>
+    <>
+      <ul className={styles.transaction}>
+        <li className={styles.transaction__item}>
+          <h4 className={styles.transaction__header}>Date</h4>
+          <p className={styles.transaction__data}>{date}</p>
+        </li>
+        <li className={styles.transaction__item}>
+          <h4 className={styles.transaction__header}>Type</h4>
+          <p className={styles.transaction__data}>
+            {income ? "Income" : "Expense"}
+          </p>
+        </li>
+        <li className={styles.transaction__item}>
+          <h4 className={styles.transaction__header}>Category</h4>
+          <p className={styles.transaction__data}>{category}</p>
+        </li>
+        <li className={styles.transaction__item}>
+          <h4 className={styles.transaction__header}>Comment</h4>
+          <p className={styles.transaction__data}>{comment}</p>
+        </li>
+        <li className={styles.transaction__item}>
+          <h4 className={styles.transaction__header}>Sum</h4>
+          <p className={styles.transaction__data}>{amount}</p>
+        </li>
+        <li className={styles.transaction__item__last}>
+          {/* TODO: add some spinner */}
+          <button className={styles.btn__delete} onClick={handleDelete}>
+            {isDeleting ? "Deleting..." : "Delete"}
+          </button>
+          <button
+            className={styles.btn__edit}
+            onClick={() => dispatch(openModal(`${_id}edit`))}
+          >
+            <Pencil className={styles.btn__edit__icon} />
+            Edit
+          </button>
+        </li>
+      </ul>
+      <EditTransaction transaction={transaction} />
+    </>
   );
 };
