+++ src/components/TransactionsListDesktopItem/TransactionsListDesktopItem.jsx
@@ -1,59 +1,46 @@
-import React, { useState } from "react";
-import { ReactComponent as Pencil } from "./../../img/pencil.svg";
 import styles from "./TransactionsListDesktopItem.module.scss";
+import React from "react";
+import { ReactComponent as Pencil } from "./../../img/pencil.svg";
+import { useDispatch } from "react-redux";
+import { openModal } from "../../redux/slices/modal/modalSlice";
 import transactionsApiSlice from "../../redux/slices/api/transactions/transactionsApiSlice";
 import EditTransaction from "../EditTransaction/EditTransaction";
 
-const TransactionsListDesktopItem = ({
-  date,
-  category,
-  income,
-  comment,
-  amount,
-  _id,
-}) => {
+const TransactionsListDesktopItem = ({ transaction }) => {
+  const { _id, date, category, comment, amount, income } = transaction;
+  const dispatch = useDispatch();
   const [deleteTransaction, { isLoading: isDeleting }] =
     transactionsApiSlice.useDeleteTransactionMutation();
 
-  const [showEdittrans, setShowEditTrans] = useState(false);
-
-  const openModal = () => {
-    setShowEditTrans(true);
-  };
-
-  const closeModal = () => {
-    setShowEditTrans(false);
+  const handleDelete = async () => {
+    await deleteTransaction(_id);
   };
 
   return (
-    <tr className={styles.transtable__row}>
-      <td className={styles.transtable__rowcell}>{date}</td>
-      <td className={styles.transtable__rowcell__type}>
-        {income ? "Income" : "Expense"}
-      </td>
-      <td className={styles.transtable__rowcell}>{category}</td>
-      <td className={styles.transtable__rowcell__com}>{comment}</td>
-      <td className={styles.transtable__rowcell__sum}>{amount}</td>
-      <td className={styles.transtable__rowcell}>
-        <button className={styles.btn__edit}>
-          <Pencil className={styles.btn__edit__icon} onClick={openModal} />
-        </button>
-        <button
-          className={styles.btn__delete}
-          onClick={async () => await deleteTransaction(_id)}
-        >
-          {/* TODO: add some spinner */}
-          {isDeleting ? "Deleting..." : "Delete"}
-        </button>
-      </td>
-      {showEdittrans && (
-        <EditTransaction
-          onClose={closeModal}
-          income={income}
-          transactionId={_id}
-        />
-      )}
-    </tr>
+    <>
+      <tr className={styles.transtable__row}>
+        <td className={styles.transtable__rowcell}>{date}</td>
+        <td className={styles.transtable__rowcell__type}>
+          {income ? "Income" : "Expense"}
+        </td>
+        <td className={styles.transtable__rowcell}>{category}</td>
+        <td className={styles.transtable__rowcell__com}>{comment}</td>
+        <td className={styles.transtable__rowcell__sum}>{amount}</td>
+        <td className={styles.transtable__rowcell}>
+          <button className={styles.btn__edit}>
+            <Pencil
+              className={styles.btn__edit__icon}
+              onClick={() => dispatch(openModal(`${_id}edit`))}
+            />
+          </button>
+          <button className={styles.btn__delete} onClick={handleDelete}>
+            {/* TODO: add some spinner */}
+            {isDeleting ? "Deleting..." : "Delete"}
+          </button>
+        </td>
+      </tr>
+      <EditTransaction transaction={transaction} />
+    </>
   );
 };
